Inline SQL namespace
