--- contrib/ruby_event_store-rom/lib/ruby_event_store/rom/adapters/sql/rake_task.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require 'rom/sql/rake_task'
-require 'ruby_event_store/rom/sql'
-load 'ruby_event_store/rom/adapters/sql/tasks/migration_tasks.rake'
