+++ contrib/ruby_event_store-rom/lib/ruby_event_store/rom/index_violation_detector.rb
