+++ contrib/ruby_event_store-rom/lib/ruby_event_store/rom/rake_task.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'rom/sql/rake_task'
+require 'ruby_event_store/rom'
+load    'ruby_event_store/rom/tasks/migration_tasks.rake'
