+++ contrib/ruby_event_store-rom/lib/ruby_event_store/rom/tasks/migration_tasks.rake
