+++ contrib/ruby_event_store-rom/spec/rom/event_repository_spec.rb
@@ -1,12 +1,11 @@
 require 'spec_helper'
-require 'ruby_event_store/rom/sql'
 require 'ruby_event_store/spec/rom/event_repository_lint'
 
 module RubyEventStore
   module ROM
     RSpec.describe EventRepository do
       include_examples :rom_event_repository
-      let(:rom_helper) { SQL::SpecHelper.new }
+      let(:rom_helper) { SpecHelper.new }
 
       def additional_limited_concurrency_for_auto_check
         positions =
