+++ contrib/ruby_event_store-rom/spec/rom/events_spec.rb
@@ -4,12 +4,10 @@ require 'ruby_event_store/spec/rom/relations/events_lint'
 
 module RubyEventStore
   module ROM
-  module SQL
-  RSpec.describe Relations::Events do
-    let(:rom_helper) { SpecHelper.new }
+    RSpec.describe Relations::Events do
+      let(:rom_helper) { SpecHelper.new }
 
-    it_behaves_like :events_relation, Relations::Events
-  end
-  end
+      it_behaves_like :events_relation, Relations::Events
+    end
   end
 end
