+++ contrib/ruby_event_store-rom/spec/rom/unit_of_work_spec.rb
@@ -1,11 +1,10 @@
 require 'spec_helper'
-require 'ruby_event_store/rom/sql'
 require 'ruby_event_store/spec/rom/unit_of_work_lint'
 
 module RubyEventStore
   module ROM
     RSpec.describe UnitOfWork do
-      let(:rom_helper) { SQL::SpecHelper.new }
+      let(:rom_helper) { SpecHelper.new }
 
       it_behaves_like :unit_of_work, UnitOfWork
     end
